% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{miss_eda}
\alias{miss_eda}
\title{Analysis of Missingness}
\usage{
miss_eda(data, na_val = NA, visual = TRUE, warn_large_data = TRUE)
}
\arguments{
\item{data}{A data.frame.}

\item{na_val}{Declared value for missings.}
}
\value{
a list of dataframes.
}
\description{
Analysis of Missingness
}
