% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{read_all_files}
\alias{read_all_files}
\title{Reads all CSV and sas7bdat files}
\usage{
read_all_files(path)
}
\arguments{
\item{path}{A path.}
}
\value{
a list of dataframes.
}
\description{
TODO:
- add Excel, ...
}
